Changed pathfinding and Scriptable Objects
Went back to an old pathfinding since I'm not sure that I'd be able to finish a pathfinding with Unity Jobs System in time but I'll take a look at it if I've got time left in the end of the project.

Also created some scriptable objects and an Boxymon script as well as redid object pooling.

Also there's some code added that I did along some lectures that'll be removed in final.